test(e2e): organize Go harness scenarios
Split suite setup and standalone scenario helpers into focused files so the Go e2e harness is easier to extend while keeping the current package boundaries.
